scripts/requirements.txt
# Function dependencies, for example:
# package>=version
requests==2.19.1
google-cloud-datastore==1.6.0
